+++ internal/goscrobble/external_lastfm.go
@@ -1,0 +1,4 @@
+package goscrobble
+
+func getImageLastFM(src string) {
+}