+++ internal/goscrobble/profile.go
@@ -1,0 +1,21 @@
+package goscrobble
+
+type ProfileResponse struct {
+	UUID      string                `json:"uuid"`
+	Username  string                `json:"username"`
+	Scrobbles []ScrobbleRequestItem `json:"scrobbles"`
+}
+
+func getProfile(user User) (ProfileResponse, error) {
+	resp := ProfileResponse{
+		UUID:     user.UUID,
+		Username: user.Username,
+	}
+	scrobbleReq, err := fetchScrobblesForUser(user.UUID, 10, 1)
+	if err != nil {
+		return resp, err
+	}
+
+	resp.Scrobbles = scrobbleReq.Items
+	return resp, nil
+}