+++ migrations/2_users.up.sql
@@ -10,6 +10,7 @@ CREATE TABLE IF NOT EXISTS `users` (
     `verified` TINYINT(1) NOT NULL DEFAULT 0,
     `active` TINYINT(1) NOT NULL DEFAULT 1,
     `admin` TINYINT(1) NOT NULL DEFAULT 0,
+    `private` TINYINT(1) NOT NULL DEFAULT 0,
     KEY `usernameLookup` (`username`, `active`),
     KEY `emailLookup` (`email`, `active`)
 ) DEFAULT CHARACTER SET = utf8mb4 COLLATE = utf8mb4_0900_ai_ci;