+++ web/src/App.js
@@ -1,9 +1,9 @@
 import { Route, Switch, withRouter } from 'react-router-dom';
-
 import Home from './Pages/Home';
 import About from './Pages/About';
 import Dashboard from './Pages/Dashboard';
 import Profile from './Pages/Profile';
+import User from './Pages/User';
 import Admin from './Pages/Admin';
 import Login from './Pages/Login';
 import Register from './Pages/Register';
@@ -14,7 +14,13 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css';
 
 const App = () => {
-    let boolTrue = true
+    let boolTrue = true;
+
+    // Remove loading spinner on load
+    const el = document.querySelector(".loader-container");
+    if (el) {
+      el.remove();
+    }
 
     return (
       <div>
@@ -24,7 +30,8 @@ const App = () => {
           <Route path="/about" component={About} />
 
           <Route path="/dashboard" component={Dashboard} />
-          <Route path="/profile" component={Profile} />
+          <Route path="/user" component={User} />
+          <Route path="/u/:uuid" component={Profile} />
 
           <Route path="/admin" component={Admin} />