+++ web/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import { HashRouter } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.min.css';
@@ -11,7 +11,7 @@ import AuthContextProvider from './Contexts/AuthContextProvider';
 
 ReactDOM.render(
   <AuthContextProvider>
-    <HashRouter>
+    <BrowserRouter>
       <ToastContainer
         position="bottom-right"
         autoClose={5000}
@@ -24,7 +24,7 @@ ReactDOM.render(
         pauseOnHover
         />
         <App />
-    </HashRouter>
+    </BrowserRouter>
   </AuthContextProvider>,
   document.getElementById('root')
 );